Use `alist-values/single-traverse` & require an equality test
The equality test must be compatible (not result in an error) with all
keys, so it doesn't make much sense to specify different equality tests
for different keys.

This simplifies `alist-let` and makes it easier to use
`alist-values/single-traverse`.

Also, now it's possible to make bindings for several occurences of the
same key.